Diep/create event v3

+++ SWP_Hackathon/src/main/java/com/hackathon/dto/criteria/CriteriaSetResponseDTO.java
@@ -13,6 +13,6 @@ import java.math.BigDecimal;
 public class CriteriaSetResponseDTO {
     private Integer criteriaSetId;
     private String criteriaName;
-    private BigDecimal maxScore;
+    private Integer maxScore;
 
 }
